ref: Extract git tasks from the packages version actions
